+++ CHANGELOG.md
@@ -26,6 +26,7 @@ omnibus-gitlab repository.
 - Added roles to ease HA Postgres configuration d7d0f32a
 - Add a runit service to probe repository storages
 - Bump bundled git version to 2.14.3 a2b4bedf
+- Render gitaly-ruby memory settings
 
 
 10.2.3